Update amazon-efs-utils.rb
Issue #, if available:

Description of changes:
Updating amazon-efs-utils to v1.34.3 in the homebrew tap and Did all the required Testing for Mounting and Unmounting and Checking the Stunnel and Mount Logs as well. It is working fine.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.